+++ examples/find_color_and_label.py
@@ -2,7 +2,8 @@ from photons_app.executor import library_setup
 from photons_app.special import FoundSerials
 
 from photons_messages import DeviceMessages, LightMessages
-from photons_control.script import Decider
+
+from collections import defaultdict
 
 collector = library_setup()
 
@@ -11,21 +12,19 @@ lan_target = collector.configuration['target_register'].resolve("lan")
 async def doit():
     getter = [DeviceMessages.GetLabel(), LightMessages.GetColor()]
 
-    def found(serial, *states):
-        info = {"label": "", "hsbk": ""}
-        for s in states:
-            if s | DeviceMessages.StateLabel:
-                info["label"] = s.label
-            elif s | LightMessages.LightState:
-                info["hsbk"] = " ".join(
-                      "{0}={1}".format(key, s.payload[key])
-                      for key in ("hue", "saturation", "brightness", "kelvin")
-                    )
-        print("{0}: {1}: {2}".format(serial, info["label"], info["hsbk"]))
-        return []
+    info = defaultdict(dict)
+    async for pkt, _, _ in lan_target.script(getter).run_with(FoundSerials()):
+        if pkt | DeviceMessages.StateLabel:
+            info[pkt.serial]["label"] = pkt.label
+        elif pkt | LightMessages.LightState:
+            hsbk = " ".join(
+                  "{0}={1}".format(key, pkt[key])
+                  for key in ("hue", "saturation", "brightness", "kelvin")
+                )
+            info[pkt.serial]["hsbk"] = hsbk
 
-    msg = Decider(getter, found, [DeviceMessages.StateLabel, LightMessages.LightState])
-    await lan_target.script(msg).run_with_all(FoundSerials())
+    for serial, details in info.items():
+        print(f"{serial}: {details['label']}: {details['hsbk']}")
 
 loop = collector.configuration["photons_app"].loop
 loop.run_until_complete(doit())